Hoan thanh xong API

+++ Helpers/FeeAPI.cs
@@ -5,21 +5,21 @@ namespace CheeseBurger.Helpers
     public class FeeAPI
     {
         [JsonProperty("total")]
-        public decimal Total { get; set; }
+        public float Total { get; set; }
 
         [JsonProperty("service_fee")]
-        public decimal ServiceFee { get; set; }
+        public int ServiceFee { get; set; }
 
         [JsonProperty("insurance_fee")]
-        public decimal InsuranceFee { get; set; }
+        public int InsuranceFee { get; set; }
 
         [JsonProperty("pick_station_fee")]
-        public decimal PickStationFee { get; set; }
+        public int PickStationFee { get; set; }
 
         [JsonProperty("coupon_value")]
-        public decimal CouponValue { get; set; }
+        public int CouponValue { get; set; }
 
         [JsonProperty("r2s_fee")]
-        public decimal R2SFee { get; set; }
+        public int R2SFee { get; set; }
     }
 }
